Refactor search result
Now that the object passed is an OpenStruct, there is no need for
single line object['field'] calls as the object exists

Replace all references to that style will object.field

Split the helper methods out so that only display methods are in the
search_result model, all other methods are stored in the helper file
